requirements.txt
bottle==0.12.25
certifi==2024.8.30
charset-normalizer==3.3.2
idna==3.8
proxy_tools==0.1.0
pycairo==1.26.1
PyGObject==3.48.2
pywebview==5.2
requests==2.32.3
typing_extensions==4.12.2
urllib3==2.2.2
